+++ src/schema/index.ts
@@ -1,4 +1,5 @@
 import { FormFieldDefinition } from './form.field.schema';
+import { FormHookDefinition } from './form.hook.schema'
 import { FormDefinition } from './form.schema';
 import { SubmissionFieldDefinition } from './submission.field.schema';
 import { SubmissionDefinition } from './submission.schema';
@@ -7,6 +8,7 @@ import { UserDefinition } from './user.schema';
 export const schema = [
   FormDefinition,
   FormFieldDefinition,
+  FormHookDefinition,
   SubmissionDefinition,
   SubmissionFieldDefinition,
   UserDefinition,