+++ src/service/submission/index.ts
@@ -3,6 +3,7 @@ import { SubmissionSetFieldService } from './submission.set.field.service';
 import { SubmissionStartService } from './submission.start.service';
 import { SubmissionStatisticService } from './submission.statistic.service';
 import { SubmissionTokenService } from './submission.token.service';
+import { SubmissionHookService } from './submission.webhook.service'
 
 export const submissionServices = [
   SubmissionService,
@@ -10,4 +11,5 @@ export const submissionServices = [
   SubmissionStartService,
   SubmissionStatisticService,
   SubmissionTokenService,
+  SubmissionHookService,
 ]